requirements.txt
# Core
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
psycopg[binary]==3.2.2
python-decouple==3.8
channels==4.1.0
passlib==1.7.4

# Admin UI theme
django-jazzmin==2.6.0

# API Schema
drf-spectacular==0.27.2

# Filtering & Pagination
django-filter==24.2

# Media
Pillow==10.4.0


# Graph DB (CDSS)
neo4j==5.19.0

# Testing
pytest==8.2.0
pytest-django==4.8.0
factory-boy==3.3.0
faker==25.0.0

# Dev tools
django-extensions==3.2.3

# Async task queue
celery==5.4.0
celery[redis]==5.4.0
redis==5.0.4
django-celery-beat==2.6.0
django-celery-results==2.5.1

# WSGI server
gunicorn==22.0.0

# Error monitoring (optional — only used when SENTRY_DSN is set)
sentry-sdk[django]==2.3.1

python-json-logger==3.3.0
